requirements.txt
Django==1.11.7
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
psycopg2==2.7.3.2
psycopg2-binary==2.8.2
pytz==2019.3
